backend/pyproject.toml
[project]
name = "geoexplorer-backend"
version = "1.0.0"
description = "GeoExplorer 地理探索学习应用后端服务"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "GeoExplorer Team" }
]

dependencies = [
    "fastapi>=0.109.2",
    "uvicorn[standard]>=0.27.1",
    "supabase>=2.3.4",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
